# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trak fitness coaching platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients/{clientId}/assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create an assignment for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /clients/{clientId}/assignments/{assignmentId}:
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{clientId}/assignments/{assignmentId}/complete:
    post:
      operationId: completeAssignment
      tags: [assignments]
      summary: Mark an assignment as complete (client action)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

  # ── Messages ──────────────────────────────────────────────────────────────
...
  /clients/{clientId}/call-logs:
    get:
      operationId: listCallLogs
      tags: [coach]
      summary: List coaching call logs for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallLog"
    post:
      operationId: createCallLog
      tags: [coach]
      summary: Log a coaching call
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallLogInput"
      responses:
        "201":
          description: Call logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
  /clients/{clientId}/call-logs/{callId}:
    patch:
      operationId: updateCallLog
      tags: [coach]
      summary: Update a call log
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: callId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallLogInput"
      responses:
        "200":
          description: Updated call log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
    delete:
      operationId: deleteCallLog
      tags: [coach]
      summary: Delete a call log
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: callId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Upload URL ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CallLog:
      type: object
      required: [id, clientId, date, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        date:
          type: string
          format: date
        durationMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CallLogInput:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date
        durationMinutes:
          type: integer
        notes:
          type: string

    # ── Measurements ────────────────────────────────────────────────────────
...
    Assignment:
      type: object
      required: [id, clientId, title, type, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [task, nutrition, mobility, habit, note]
        body:
          type: ["string", "null"]
        targetValue:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [pending, completed]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    AssignmentInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [task, nutrition, mobility, habit, note]
        body:
          type: string
        targetValue:
          type: string
        dueDate:
          type: string
          format: date
    AssignmentUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [task, nutrition, mobility, habit, note]
        body:
          type: ["string", "null"]
        targetValue:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [pending, completed]

    # ── Messages ────────────────────────────────────────────────────────────
...